move button style config to widgetry
