+++ map_gui/src/colors.rs
@@ -53,15 +53,6 @@ impl ColorSchemeChoice {
     }
 }
 
-pub struct ButtonColorScheme {
-    pub fg: Color,
-    pub fg_disabled: Color,
-    pub outline: Color,
-    pub bg: Color,
-    pub bg_hover: Color,
-    pub bg_disabled: Color,
-}
-
 pub struct ColorScheme {
     scheme: ColorSchemeChoice,
 
@@ -86,10 +77,6 @@ pub struct ColorScheme {
     pub minimap_unselected_zoom: Color,
 
     // UI > Buttons
-    pub btn_primary_dark: ButtonColorScheme,
-    pub btn_secondary_dark: ButtonColorScheme,
-    pub btn_primary_light: ButtonColorScheme,
-    pub btn_secondary_light: ButtonColorScheme,
 
     // Roads
     driving_lane: Color,
@@ -217,40 +204,6 @@ impl ColorScheme {
             minimap_selected_zoom: Color::WHITE,
             minimap_unselected_zoom: Color::WHITE.alpha(0.2),
             gui_style,
-
-            // UI > Buttons
-            btn_primary_dark: ButtonColorScheme {
-                fg: hex("#4C4C4C"),
-                fg_disabled: hex("#4C4C4C").alpha(0.3),
-                bg: Color::WHITE.alpha(0.8),
-                bg_hover: Color::WHITE,
-                bg_disabled: Color::grey(0.6),
-                outline: Color::WHITE.alpha(0.6),
-            },
-            btn_secondary_dark: ButtonColorScheme {
-                fg: hex("#4C4C4C"),
-                fg_disabled: hex("#4C4C4C").alpha(0.3),
-                bg: Color::CLEAR,
-                bg_hover: hex("#4C4C4C").alpha(0.1),
-                bg_disabled: Color::grey(0.8),
-                outline: hex("#4C4C4C"),
-            },
-            btn_primary_light: ButtonColorScheme {
-                fg: hex("#F2F2F2"),
-                fg_disabled: hex("#F2F2F2").alpha(0.3),
-                bg: hex("#003046").alpha(0.8),
-                bg_hover: hex("#003046"),
-                bg_disabled: Color::grey(0.1),
-                outline: hex("#003046").alpha(0.6),
-            },
-            btn_secondary_light: ButtonColorScheme {
-                fg: hex("#F2F2F2"),
-                fg_disabled: hex("#F2F2F2").alpha(0.3),
-                bg: Color::CLEAR,
-                bg_hover: hex("#F2F2F2").alpha(0.1),
-                bg_disabled: Color::grey(0.9),
-                outline: hex("#F2F2F2"),
-            },
 
             // Roads
             driving_lane: Color::BLACK,
