+++ widgetry/src/widgets/persistent_split.rs
@@ -73,7 +73,7 @@ impl<T: 'static + PartialEq + Clone + std::fmt::Debug> PersistentSplit<T> {
 // I'd like to base it on ColorScheme, but that currently lives in map_gui, so for now
 // I've hardcoded the builder colors. We literally use it in one place.
 fn button_builder<'a>() -> ButtonBuilder<'a> {
-    // let primary_light = ButtonColorScheme {
+    // let primary_light = ButtonStyle {
     //     fg: hex("#F2F2F2"),
     //     fg_disabled: hex("#F2F2F2").alpha(0.3),
     //     bg: hex("#003046").alpha(0.8),
